requirements.txt
pandas
joblib==1.5.2
numpy
scikit-learn==1.6.1
kagglehub
nltk
matplotlib
seaborn
xgboost
lightgbm
python-dotenv
mlflow==3.3.0
azure-storage-blob
IPython
azure-identity
sqlalchemy
psycopg2-binary
nltk
locust
redis